phy/lpddr4: extract common test helpers for use when testing other PHYs
